refactor: OperationTimeChecker 클래스명 및 구조 변경
