+++ src/main/java/domain/OperationTime.java
@@ -1,0 +1,26 @@
+package domain;
+
+import java.time.DayOfWeek;
+import java.time.LocalDate;
+import java.time.LocalTime;
+
+public class OperationTime {
+
+    public static boolean isContainsOperationTime(final LocalTime targetTime) {
+        return !targetTime.isBefore(LocalTime.of(8, 0)) && !targetTime.isAfter(LocalTime.of(23, 0));
+    }
+
+    public static void checkIsOperationDate(final LocalDate targetDate) {
+        if (isWeekend(targetDate) || isHoliday(targetDate)) {
+            throw new IllegalArgumentException("운영일이 아닙니다.");
+        }
+    }
+
+    private static boolean isWeekend(final LocalDate targetDate) {
+        return targetDate.getDayOfWeek() == DayOfWeek.SUNDAY || targetDate.getDayOfWeek() == DayOfWeek.SATURDAY;
+    }
+
+    private static boolean isHoliday(final LocalDate targetDate) {
+        return targetDate.isEqual(LocalDate.of(2024, 12, 25));
+    }
+}
